add dependency `javax.annotation:jaax.annotation-api` to NoClassDefFoundError: org/apache/pulsar/shade/javax/annotation/Priority.
```
Caused by: java.lang.NoClassDefFoundError: org/apache/pulsar/shade/javax/annotation/Priority
at org.apache.pulsar.shade.org.glassfish.jersey.model.internal.CommonConfig$FeatureRegistration.priority(CommonCon {}
at org.apache.pulsar.shade.org.glassfish.jersey.model.internal.CommonConfig$FeatureRegistration.<init>(CommonConfig.java:119)
at org.apache.pulsar.shade.org.glassfish.jersey.model.internal.CommonConfig$FeatureRegistration.<init>(CommonConfig.java:107)
at org.apache.pulsar.shade.org.glassfish.jersey.model.internal.CommonConfig.processFeatureRegistration(CommonConfig.java:511)
at org.apache.pulsar.shade.org.glassfish.jersey.model.internal.CommonConfig.register(CommonConfig.java:418)
at org.apache.pulsar.shade.org.glassfish.jersey.client.ClientConfig$State.register(ClientConfig.java:216)
at org.apache.pulsar.shade.org.glassfish.jersey.client.ClientConfig.register(ClientConfig.java:606)
at org.apache.pulsar.client.admin.internal.PulsarAdminImpl.<init>(PulsarAdminImpl.java:179)
at org.apache.pulsar.client.admin.internal.PulsarAdminBuilderImpl.build(PulsarAdminBuilderImpl.java:47)
at org.apache.pulsar.sql.presto.PulsarConnectorConfig.getPulsarAdmin(PulsarConnectorConfig.java:474)
... 51 more {}
Caused by: java.lang.ClassNotFoundException: org.apache.pulsar.shade.javax.annotation.Priority
at java.base/java.net.URLClassLoader.findClass(URLClassLoader.java:445)
at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:587)
at io.prestosql.server.PluginClassLoader.loadClass(PluginClassLoader.java:89)
at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:520)
... 61 more
```
